941 add navigation in popup
