+++ frontend/src/components/MapView/LeftPanel/ChartsPanel/index.tsx
@@ -17,10 +17,8 @@ import {
   Typography,
 } from '@material-ui/core';
 import { GeoJsonProperties } from 'geojson';
-import { groupBy, mapKeys, snakeCase } from 'lodash';
 import React, {
   memo,
-  MutableRefObject,
   useCallback,
   useEffect,
   useMemo,
@@ -39,8 +37,8 @@ import { LayerData } from 'context/layers/layer-data';
 import { leftPanelTabValueSelector } from 'context/leftPanelStateSlice';
 import { layerDataSelector } from 'context/mapStateSlice/selectors';
 import { useSafeTranslation } from 'i18n';
-import { castObjectsArrayToCsv } from 'utils/csv-utils';
-import { downloadToFile } from 'components/MapView/utils';
+import DownloadCsvButton from 'components/MapView/DownloadCsvButton';
+
 import ChartSection from './ChartSection';
 import LocationSelector from './LocationSelector';
 import TimePeriodSelector from './TimePeriodSelector';
@@ -114,17 +112,6 @@ const useStyles = makeStyles(() =>
       marginTop: 0,
       paddingBottom: '1em',
     },
-    downloadButton: {
-      backgroundColor: '#62B2BD',
-      '&:hover': {
-        backgroundColor: '#62B2BD',
-      },
-      marginTop: '2em',
-      marginLeft: '25%',
-      marginRight: '25%',
-      width: '50%',
-      '&.Mui-disabled': { opacity: 0.5 },
-    },
     clearAllSelectionsButton: {
       backgroundColor: '#788489',
       '&:hover': {
@@ -177,81 +164,6 @@ const menuProps: Partial<MenuProps> = {
       width: 'auto',
     },
   },
-};
-
-// We export the downloadCsv function to be tested independently
-export const downloadCsv = (
-  params: [MutableRefObject<{ [key: string]: any[] }>, string][],
-  // filename1: string,
-  // dataForSecondCsv: MutableRefObject<{ [key: string]: any[] }>,
-  // filename2: string,
-) => {
-  return () => {
-    params.forEach(filedata => {
-      const [dataForCsv, filename] = filedata;
-
-      const dateColumn = 'Date';
-      const getKeyName = (key: string, chartName: string) =>
-        key.endsWith('_avg')
-          ? `${snakeCase(chartName)}_avg`
-          : snakeCase(chartName);
-
-      const columnsNamesPerChart = Object.entries(dataForCsv.current).map(
-        ([key, value]) => {
-          const first = value[0];
-          const keys = Object.keys(first);
-          const filtered = keys.filter(x => x !== dateColumn);
-          const mapped = filtered.map(x => getKeyName(x, key));
-          return Object.fromEntries(mapped.map(x => [x, x]));
-        },
-      );
-
-      const columnsNames = columnsNamesPerChart.reduce(
-        (prev, curr) => ({ ...prev, ...curr }),
-        { [dateColumn]: dateColumn },
-      );
-
-      const merged = Object.entries(dataForCsv.current)
-        .map(([key, value]) => {
-          return value.map(x => {
-            return mapKeys(x, (v, k) =>
-              k === dateColumn ? dateColumn : getKeyName(k, key),
-            );
-          });
-        })
-        .flat();
-      if (merged.length < 1) {
-        return;
-      }
-
-      const grouped = groupBy(merged, dateColumn);
-      // The blueprint of objects array data
-      const initialObjectsArrayBlueprintData = Object.keys(columnsNames).reduce(
-        (acc: { [key: string]: string }, key) => {
-          // eslint-disable-next-line fp/no-mutation
-          acc[key] = '';
-          return acc;
-        },
-        {},
-      );
-
-      const objectsArray = Object.entries(grouped).map(([, value]) => {
-        return value.reduce(
-          (prev, curr) => ({ ...prev, ...curr }),
-          initialObjectsArrayBlueprintData,
-        );
-      });
-
-      downloadToFile(
-        {
-          content: castObjectsArrayToCsv(objectsArray, columnsNames, ','),
-          isUrl: false,
-        },
-        filename,
-        'text/csv',
-      );
-    });
-  };
 };
 
 const ChartsPanel = memo(
@@ -318,55 +230,6 @@ const ChartsPanel = memo(
     const { t } = useSafeTranslation();
 
     const tabValue = useSelector(leftPanelTabValueSelector);
-
-    const generateCSVFilename = useCallback(() => {
-      return [
-        multiCountry ? admin0Key : country,
-        selectedAdmin1Area ?? '',
-        selectedAdmin2Area ?? '',
-        ...selectedLayerTitles,
-        comparePeriods ? 'first_period' : '',
-      ]
-        .filter(x => !!x)
-        .map(snakeCase)
-        .join('_');
-    }, [
-      admin0Key,
-      comparePeriods,
-      country,
-      multiCountry,
-      selectedAdmin1Area,
-      selectedAdmin2Area,
-      selectedLayerTitles,
-    ]);
-
-    const generateSecondCSVFilename = useCallback(() => {
-      return [
-        multiCountry ? secondAdmin0Key : country,
-        compareLocations
-          ? secondSelectedAdmin1Area ?? ''
-          : selectedAdmin1Area ?? '',
-        compareLocations
-          ? secondSelectedAdmin2Area ?? ''
-          : selectedAdmin2Area ?? '',
-        ...selectedLayerTitles,
-        comparePeriods ? 'second_period' : '',
-      ]
-        .filter(x => !!x)
-        .map(snakeCase)
-        .join('_');
-    }, [
-      country,
-      compareLocations,
-      multiCountry,
-      secondSelectedAdmin1Area,
-      selectedAdmin1Area,
-      secondAdmin0Key,
-      secondSelectedAdmin2Area,
-      selectedAdmin2Area,
-      selectedLayerTitles,
-      comparePeriods,
-    ]);
 
     const onChangeChartLayers = useCallback(
       (event: React.ChangeEvent<{ value: unknown }>) => {
@@ -887,12 +750,18 @@ const ChartsPanel = memo(
             ))}
           </Select>
         </FormControl>
-        <Button
-          className={classes.downloadButton}
-          onClick={downloadCsv([
-            [dataForCsv, generateCSVFilename()],
-            [dataForSecondCsv, generateSecondCSVFilename()],
-          ])}
+        <DownloadCsvButton
+          admin0Key={admin0Key}
+          compareLocations={compareLocations}
+          comparePeriods={comparePeriods}
+          country={country}
+          multiCountry={multiCountry}
+          secondAdmin0Key={secondAdmin0Key}
+          secondSelectedAdmin1Area={secondSelectedAdmin1Area}
+          secondSelectedAdmin2Area={secondSelectedAdmin2Area}
+          selectedAdmin1Area={selectedAdmin1Area}
+          selectedAdmin2Area={selectedAdmin2Area}
+          selectedLayerTitles={selectedLayerTitles}
           disabled={
             !(
               adminProperties &&
@@ -901,9 +770,9 @@ const ChartsPanel = memo(
               selectedLayerTitles.length >= 1
             )
           }
-        >
-          <Typography variant="body2">{t('Download CSV')}</Typography>
-        </Button>
+          dataForCsv={dataForCsv}
+          dataForSecondCsv={dataForSecondCsv}
+        />
         <Button
           className={classes.clearAllSelectionsButton}
           onClick={handleClearAllSelectedCharts}
