+++ frontend/src/components/MapView/utils.ts
@@ -1,4 +1,4 @@
-import { orderBy, values } from 'lodash';
+import { groupBy, mapKeys, orderBy, snakeCase, values } from 'lodash';
 import { Map } from 'mapbox-gl';
 import { TFunction } from 'i18next';
 import { Dispatch } from 'redux';
@@ -18,6 +18,9 @@ import { addNotification } from 'context/notificationStateSlice';
 import { LocalError } from 'utils/error-utils';
 import { Column, quoteAndEscapeCell } from 'utils/analysis-utils';
 import { TableRow } from 'context/analysisResultStateSlice';
+import { MutableRefObject } from 'react';
+import { castObjectsArrayToCsv } from 'utils/csv-utils';
+
 import { getExtent } from './Layers/raster-utils';
 
 export const getActiveFeatureInfoLayers = (map: Map): WMSLayerProps[] => {
@@ -85,6 +88,87 @@ export const downloadToFile = (
   );
   link.setAttribute('download', `${filename}.${fileType}`);
   link.click();
+};
+
+export const downloadCsv = (
+  params: [MutableRefObject<{ [key: string]: any[] }>, string][],
+  // filename1: string,
+  // dataForSecondCsv: MutableRefObject<{ [key: string]: any[] }>,
+  // filename2: string,
+) => {
+  return () => {
+    params.forEach(filedata => {
+      const [dataForCsv, filename] = filedata;
+
+      const dateColumn = 'Date';
+      const getKeyName = (key: string, chartName: string) =>
+        key.endsWith('_avg')
+          ? `${snakeCase(chartName)}_avg`
+          : snakeCase(chartName);
+
+      const columnsNamesPerChart = Object.entries(dataForCsv.current).map(
+        ([key, value]) => {
+          const first = value[0];
+          const keys = Object.keys(first);
+          const filtered = keys.filter(x => x !== dateColumn);
+          const mapped = filtered.map(x => getKeyName(x, key));
+          return Object.fromEntries(mapped.map(x => [x, x]));
+        },
+      );
+
+      const columnsNames = columnsNamesPerChart.reduce(
+        (prev, curr) => ({ ...prev, ...curr }),
+        { [dateColumn]: dateColumn },
+      );
+
+      const merged = Object.entries(dataForCsv.current)
+        .map(([key, value]) => {
+          return value.map(x => {
+            return mapKeys(x, (v, k) =>
+              k === dateColumn ? dateColumn : getKeyName(k, key),
+            );
+          });
+        })
+        .flat();
+      if (merged.length < 1) {
+        return;
+      }
+
+      const grouped = groupBy(merged, dateColumn);
+      // The blueprint of objects array data
+      const initialObjectsArrayBlueprintData = Object.keys(columnsNames).reduce(
+        (acc: { [key: string]: string }, key) => {
+          // eslint-disable-next-line fp/no-mutation
+          acc[key] = '';
+          return acc;
+        },
+        {},
+      );
+
+      const objectsArray = Object.entries(grouped).map(([, value]) => {
+        return value.reduce(
+          (prev, curr) => ({ ...prev, ...curr }),
+          initialObjectsArrayBlueprintData,
+        );
+      });
+
+      downloadToFile(
+        {
+          content: castObjectsArrayToCsv(objectsArray, columnsNames, ','),
+          isUrl: false,
+        },
+        filename,
+        'text/csv',
+      );
+    });
+  };
+};
+
+export const buildCsvFileName = (items: string[]) => {
+  return items
+    .filter(x => !!x)
+    .map(snakeCase)
+    .join('_');
 };
 
 export function getFeatureInfoPropsData(
